Extract `mesh_pipeline_layout` file
