+++ crates/bevy_pbr/src/render/mod.rs
@@ -2,6 +2,7 @@ mod fog;
 mod light;
 pub(crate) mod mesh;
 mod mesh_bindings;
+mod mesh_pipeline_layout;
 mod morph;
 mod skin;
 
@@ -9,4 +10,5 @@ pub use fog::*;
 pub use light::*;
 pub use mesh::*;
 pub use mesh_bindings::MeshLayouts;
+pub use mesh_pipeline_layout::*;
 pub use skin::{extract_skins, prepare_skins, SkinIndex, SkinUniform, MAX_JOINTS};
